requirements.txt
pyusb
Pillow
audioread
fs